Add code refactoring and minor UI changes
