+++ ui/components/app/app-components.scss
@@ -43,7 +43,6 @@
 @import 'flask/snap-settings-card/index';
 @import 'flask/update-snap-permission-list/index';
 @import 'flask/copyable/index';
-@import "flask/permission-cell/index";
 @import 'gas-details-item/index';
 @import 'gas-details-item/gas-details-item-title/index';
 @import 'gas-timing/index';
@@ -60,6 +59,7 @@
 @import 'permissions-connect-footer/index';
 @import 'permissions-connect-header/index';
 @import 'permissions-connect-permission-list/index';
+@import 'permission-cell/index';
 @import 'recovery-phrase-reminder/index';
 @import 'set-approval-for-all-warning/index';
 @import 'step-progress-bar/index.scss';
