+++ ui/components/app/permission-cell/index.scss
