+++ ui/components/app/permissions-connect-permission-list/permissions-connect-permission-list.js
@@ -2,13 +2,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { getWeightedPermissions } from '../../../helpers/utils/permission';
 import { useI18nContext } from '../../../hooks/useI18nContext';
-import PermissionCell from '../flask/permission-cell';
+import PermissionCell from '../permission-cell';
+import Box from '../../ui/box';
 
 export default function PermissionsConnectPermissionList({ permissions }) {
   const t = useI18nContext();
 
   return (
-    <div className="permissions-connect-permission-list">
+    <Box paddingTop={1}>
       {getWeightedPermissions(t, permissions).map((perm, index) => {
         return (
           <PermissionCell
@@ -21,7 +22,7 @@ export default function PermissionsConnectPermissionList({ permissions }) {
           />
         );
       })}
-    </div>
+    </Box>
   );
 }
