+++ ui/pages/permissions-connect/flask/snap-install/index.scss
@@ -4,14 +4,6 @@
   .headers {
     flex: 1;
 
-    .permissions-connect-permission-list {
-      padding: 0 24px;
-
-      .permission {
-        padding: 8px 0;
-      }
-    }
-
     .loader-container {
       height: 100%;
     }
@@ -20,7 +12,6 @@
       border-bottom: 1px solid var(--color-border-default);
     }
   }
-
 
   .page-container__footer {
     width: 100%;
